Change the type of height and width in Render to float.
This is for compatibility with 6.10 and prior versions of Ren'Py, and
because it makes sense. Zooming an image wants to create an image with
a non-integer width and height. This is especially true when xalign is
non-zero - we need to take the precise width into account when
centering the displayable (at least when using subpixel precision).

+++ renpy/display/render.pxd
@@ -10,7 +10,7 @@ cdef class Render:
 
     cdef public bint mark, cache_killed
 
-    cdef public int width, height
+    cdef public float width, height
     cdef public object layer_name
 
     cdef public list children
